Add input/output atten, CombSweep type and reorganize:
